% Generated by roxygen2 (4.0.1): do not edit by hand
\name{A_E}
\alias{A_E}
\title{A_E Function, Automatic Edges}
\usage{
A_E(DATA, Nmin, Nmax)
}
\arguments{
\item{DATA}{}

\item{Nmin}{}

\item{Nmax}{}
}
\value{
edges
}
\description{
Probably the funciton to set automatic edges for the processing windows
}

