% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_data.R
\name{align_data}
\alias{align_data}
\title{Alignment of GC/MS chromatograms in \code{GCMS} package}
\usage{
align_data(projectpath)
}
\arguments{
\item{projectpath}{base path from \code{GCMS} processing script}
}
\value{
a directory \code{alignement} is created that includes the
related to chromatogram alignment.
}
\description{
Chromatogram alignment by maximizing the covariance between the
ion-channels or total ion counts (TIC) of the individual chromatograms.
}
\details{
creates a number of files:
\itemize{
\item{A_DATA}{list with 'TIC','ic','bp', 'ok', 'ion'}
\item{SUM_MZ}{sum of ion counts per mz}
\item{COLUMNID1}{filenames without extension}
\item{SCAN_RANGE}{sequence of all mz values, is obtained from rawimport files}
\item{NUM_scans}{nrow Xbc, how many data points}
}
}
\author{
Lorenz Gerber
}
