% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_samples.R
\name{aligndatapred}
\alias{aligndatapred}
\title{function aligndatapred}
\usage{
aligndatapred(projectpath, predpath, newsamples, target, minshift, datasource,
  ion)
}
\description{
Function aligndatapred
}

