% Generated by roxygen2 (4.0.1): do not edit by hand
\name{approxExtrap}
\alias{approxExtrap}
\title{Function approxExtrap}
\usage{
approxExtrap(x, y, xout, method = "linear", n = 50, rule = 2, f = 0,
  ties = "ordered", na.rm = FALSE)
}
\arguments{
\item{x}{}

\item{y}{}

\item{xout}{}

\item{method}{}

\item{n}{}

\item{rule}{}

\item{f}{}

\item{ties}{}

\item{na.rm}{}
}
\value{
list(x,y)
}
\description{
Function approxExtrap
This function is called from find_spectrum.R, find_spectrum2.R, spec2NIST.R and spec2NIST2.R
}

