% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAgilent.R, R/read_cdf.R
\name{baseline}
\alias{baseline}
\alias{baseline}
\title{Function Baseline}
\usage{
baseline(X, projectpath, FL = 0)

baseline(X, projectpath, FL = 0)
}
\arguments{
\item{X}{raw imported Data. Time as rows and mz as columns}

\item{projectpath}{baseproject directory}

\item{FL}{Filterlength}

\item{X}{raw imported Data. Time as rows and mz as columns}

\item{projectpath}{baseproject directory}

\item{FL}{Filterlength}
}
\value{
Xbc           moving average filtered data table, one file

Xbc           moving average filtered data table, one file
}
\description{
Function Baseline

Function Baseline
}
