% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfRewrite.R
\name{cdfRewrite}
\alias{cdfRewrite}
\title{Function cdfRewrite}
\usage{
cdfRewrite(projectpath, filepath)
}
\arguments{
\item{filepath}{can be used to rewrite files in batchmode}
}
\description{
Function cdfRewrite
}

