% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_AR_all}
\alias{do_AR_all}
\title{Function do_AR_all}
\usage{
do_AR_all(x, projectpath, NL, RP, RT_LIMIT, SCAN_RANGE)
}
\arguments{
\item{x}{}

\item{projectpath}{}

\item{NL}{}

\item{RP}{}

\item{RT_LIMIT}{}

\item{SCAN_RANGE}{}
}
\value{
C, S, INDEX, R2, noise, CP
}
\description{
Function do_AR_all
Funciton to do Alternate Regression
}

