% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_AR_all_prediction}
\alias{do_AR_all_prediction}
\title{Function do_AR_all_prediction}
\usage{
do_AR_all_prediction(x, S, CP, RT_LIMIT)
}
\arguments{
\item{x}{}

\item{S}{}

\item{CP}{}

\item{RT_LIMIT}{}
}
\value{
Cnew
}
\description{
Function do_AR_all_prediction
Funciton to do Alternate Regression on a prediction set
}

