% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cdf.R
\name{do_DATA1}
\alias{do_DATA1}
\title{Function do_DATA1}
\usage{
do_DATA1(INT, MZ, TIME, count, MZmin, MZmax, MZP)
}
\arguments{
\item{INT}{intensity values from CDF import}

\item{MZ}{mass values from CDF import}

\item{TIME}{scantime values from CDF import}

\item{count}{point count from CDF import}

\item{MZmin}{min value mass range from CDF import}

\item{MZmax}{max value mass range from CDF import}

\item{MZP}{mass precision from settings file}
}
\value{
DATA  rawdata matrix with rows for timepoints
and columns for mz values. The mz values range from 
zero to maxMZ.
}
\description{
Function do_DATA1 is the method used with Agilent Chemstation data 
and is set as default. It is called from read_cdf.R
This function is used to arrange the initial data table after import
}
