% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_DATA2}
\alias{do_DATA2}
\title{Function do_DATA2}
\usage{
do_DATA2(INT, MZ, TIME, count, MZmin, MZmax, MZP)
}
\arguments{
\item{INT}{}

\item{MZ}{}

\item{TIME}{}

\item{count}{}

\item{MZmin}{}

\item{MZmax}{}

\item{MZP}{}
}
\value{
DATA
}
\description{
Function do_DATA2
}

