% Generated by roxygen2 (4.0.1): do not edit by hand
\name{find_spectrum}
\alias{find_spectrum}
\title{Funcion find_spectrum}
\usage{
find_spectrum(projectpath, incl, extra, windowlist)
}
\arguments{
\item{projectpath}{}

\item{incl}{}

\item{extra}{}

\item{windowlist}{}
}
\value{
windowlist
}
\description{
Function find_spectrum
}

