% Generated by roxygen2 (4.0.1): do not edit by hand
\name{find_spectrum2}
\alias{find_spectrum2}
\title{Function find_spectrum2}
\usage{
find_spectrum2(predpath, projectpath)
}
\arguments{
\item{predpath}{}

\item{projectpath}{}
}
\value{
type, windwolist
}
\description{
Function find_spectrum2
}

