% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCMS-package.R
\docType{package}
\name{GCMS-package}
\alias{GCMS}
\alias{GCMS-package}
\title{GC/MS Data Processing}
\description{
GC/MS Data Processing Pipeline
}
\details{
\tabular{ll}{
Package: \tab GCMS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 25-01-2012\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes \cr
}

The package allows processing of Gas Chromatography/Mass-Spectrometry data
available in NetCDF format. It includes import, smoothing, and background
correction of chromatograms aswell as peak deconvolution, and peak integration.
The peak detection is based on Multivariate Curve Resolution by Alternate
Regresstion (MCR-ALS) as shown in Jonsson et al. (2005).
}
\author{
Lorenz Gerber \email{lorenz.gerber@slu.se}
}
\keyword{package}

