% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCR_settings.R, R/check_data.R, R/new_samples.R
\name{pca}
\alias{pca}
\title{Function pca}
\usage{
pca(X, comp)

pca(X, comp)

pca(X, comp)
}
\value{
vec, p

vec, p

vec, p
}
\description{
Function pca

Function pca

Function pca
}

