% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_proc.R, R/new_samples.R
\name{peak_pick}
\alias{peak_pick}
\alias{peak_pick}
\title{Function peak_pick}
\usage{
peak_pick(x)

peak_pick(x)
}
\value{
xpeak, xout

xpeak, xout
}
\description{
Function peak_pick
Function to find peaks

Function peak_pick
Function to find peaks
}
