% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_proc.R
\name{pure}
\alias{pure}
\title{Function pure}
\usage{
pure(d, nr, f)
}
\value{
sp, imp
}
\description{
Function pure
[sp,imp]=pure(d,nr,f)
sp purest row/column profiles
imp indexes of purest variables
d data matrix; nr (rank) number of pure components to search
if d(nspectra,nwave) imp gives purest nwave => sp are conc. profiles (nr,nspectra)
if d(nwave,nspectra) imp gives purest nspectra => sp are spectra profiles (nr,nwave)
f percent of noise allowed respect maximum of the average spectrum given in % (i.e. 1% or 0.1%))
}
