% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cdf.R
\name{read_cdf}
\alias{read_cdf}
\title{Function read_cdf}
\usage{
read_cdf(projectpath, filepath)
}
\arguments{
\item{filepath}{can be used to import a single file and return the data to the caller}
}
\value{
Xbc, SCAN_INFO, SCAN_RANGE, file
}
\description{
Function read_cdf
}
