% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_win2}
\alias{read_win2}
\title{Function read_win2}
\usage{
read_win2(projectpath, type, win)
}
\arguments{
\item{projectpath}{}

\item{type}{}

\item{win}{}
}
\value{
DATA, SPECTRUM, VARID1, OBSID1, VARID2
}
\description{
Function read_win2
}

