% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAgilent.R
\name{readAgilent}
\alias{readAgilent}
\title{Function readAgilent}
\usage{
readAgilent(projectpath, filepath)
}
\arguments{
\item{filepath}{can be used to import a single file and return the data to the caller}
}
\value{
Xbc, SCAN_INFO, SCAN_RANGE, file
}
\description{
Function readAgilent
}

