% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAgilent.R
\name{readDFile}
\alias{readDFile}
\title{Function readDFile}
\usage{
readDFile(pathname)
}
\arguments{
\item{pathname}{the pathname of the directory containing the data to import}
}
\value{
outData       list containing importInt, importMz, importCount, importScanTime
}
\description{
Function readDFile
}

