% Generated by roxygen2 (4.0.1): do not edit by hand
\name{replot}
\alias{replot}
\title{Function replot}
\usage{
replot(plotlimits = FALSE, plotedges = FALSE, plotfigure = FALSE, ymax,
  edges, DATA, num, zoomwidth, maintext, min_win, max_win)
}
\arguments{
\item{plotlimits}{}

\item{plotedges}{}

\item{plotfigure}{}

\item{ymax}{}

\item{edges}{}

\item{DATA}{}

\item{num}{}

\item{zoomwidth}{}

\item{maintext}{}

\item{min_win}{}

\item{max_win}{}
}
\description{
Function replot
}

