% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_data.R
\name{replot_align}
\alias{replot_align}
\title{Function replot_align}
\usage{
replot_align(DATA, ADATA, start, stop, zoomwidth = c(0, 0), datasource,
  maintext, targetfile)
}
\arguments{
\item{DATA}{the target data}

\item{ADATA}{the shifted data}

\item{start}{start position for the chromatograms}

\item{stop}{stop position for the chromatograms}

\item{zoomwidth}{zoomwidth}
}
\description{
Function replot_align
}

