% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{replot_checkdata}
\alias{replot_checkdata}
\title{Function replot_checkdata}
\usage{
replot_checkdata(x, y, scale_text, maintext, OBS, ssX, select, coords = NULL)
}
\description{
Function replot_checkdata
}

