% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCR_settings.R
\name{replot_mcr}
\alias{replot_mcr}
\title{Function replot_mcr}
\usage{
replot_mcr(x, y, maintext, incl, excl, numfiles, set1, set2,
  seltype = (!missing(set1) & !missing(set2)), plotlabels = TRUE, pred)
}
\arguments{
\item{pred}{}
}
\description{
Function replot_mcr
}
