% Generated by roxygen2 (4.0.1): do not edit by hand
\name{separate_windows}
\alias{separate_windows}
\title{Function separate_windows}
\usage{
separate_windows(projectpath, windowlist, files, edges, model, prediction)
}
\arguments{
\item{projectpath}{}

\item{windowlist}{}

\item{files}{}

\item{edges}{}

\item{model}{}

\item{prediction}{}
}
\description{
Function separate_windows
}

