% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{settings}
\alias{settings}
\title{GUI for general GCMS Data processing settings}
\usage{
settings(projectpath, setdefault = FALSE)
}
\arguments{
\item{projectpath}{Projectpath where the directory structure is built up in}

\item{setdefault}{If \code{TRUE}, all values are set to default.}
}
\value{
none
}
\description{
Basic Settings for GCMS Data Processing
}
\details{
\code{settings} creates the file \code{settings.Rdata} in the current
workdirectory with default settings for the \code{GCMS} data processing
script. If the file already exists a console based user interface to
modify the values.
}
\author{
Lorenz Gerber \email{lorenz.gerber@slu.se}
}

