% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_proc.R, R/gui_set_edges.R, R/new_samples.R
\name{sgolay}
\alias{sgolay}
\title{Function sgolay}
\usage{
sgolay(p, n, m = 0, ts = 1)

sgolay(p, n, m = 0, ts = 1)

sgolay(p, n, m = 0, ts = 1)
}
\value{
F

F

F
}
\description{
Function sgolay

Function sgolay

Function sgolay
}

