% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_proc.R, R/gui_set_edges.R, R/new_samples.R
\name{sgolayfilt}
\alias{sgolayfilt}
\title{Function sgolayfilt}
\usage{
sgolayfilt(x, p = 3, n = p + 3 - p\%\%2, m = 0, ts = 1)

sgolayfilt(x, p = 3, n = p + 3 - p\%\%2, m = 0, ts = 1)

sgolayfilt(x, p = 3, n = p + 3 - p\%\%2, m = 0, ts = 1)
}
\description{
Function sgolayfilt

Function sgolayfilt

Function sgolayfilt
}

