% Generated by roxygen2 (4.0.1): do not edit by hand
\name{align_data}
\alias{align_data}
\title{Alignment of GC/MS chromatograms in \code{GCMS} package}
\usage{
align_data(projectpath)
}
\arguments{
\item{projectpath}{base path from \code{GCMS} processing script}
}
\value{
a directory \code{alignement} is created that includes the
related to chromatogram alignment.
}
\description{
Chromatogram alignment by maximizing the covariance between the
ion-channels or total ion counts (TIC) of the individual chromatograms.
}
\details{
creates a number of files:
\itemize{
\item{A_DATA}{list with 'TIC','ic','bp'}
\item{ADATA}{list with 'TIC','ic','bp'}
}
}
\author{
Lorenz Gerber
}

