% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aligndatapred}
\alias{aligndatapred}
\title{function aligndatapred}
\usage{
aligndatapred(projectpath, predpath, newsamples, target, minshift, datasource,
  ion)
}
\arguments{
\item{projectpath}{}

\item{predpath}{}

\item{newsamples}{}

\item{target}{}

\item{minshift}{}

\item{datasource}{}

\item{ion}{}
}
\description{
Function aligndatapred
}

