% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec2NIST.R, R/spec2NIST2.R
\name{approxExtrap}
\alias{approxExtrap}
\title{Function approxExtrap}
\usage{
approxExtrap(x, y, xout, method = "linear", n = 50, rule = 2, f = 0,
  ties = "ordered", na.rm = FALSE)

approxExtrap(x, y, xout, method = "linear", n = 50, rule = 2, f = 0,
  ties = "ordered", na.rm = FALSE)
}
\value{
list(x,y)

list(x,y)
}
\description{
Function approxExtrap
This function is called from find_spectrum.R, find_spectrum2.R, spec2NIST.R and spec2NIST2.R

Function approxExtrap
This function is called from find_spectrum.R, find_spectrum2.R, spec2NIST.R and spec2NIST2.R
}

