% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_proc.R
\name{do_AR_all}
\alias{do_AR_all}
\title{Function do_AR_all}
\usage{
do_AR_all(x, projectpath, NL, RP, RT_LIMIT, SCAN_RANGE)
}
\value{
C, S, INDEX, R2, noise, CP
}
\description{
Function do_AR_all
Funciton to do Alternate Regression
}

