% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_DATA1}
\alias{do_DATA1}
\title{Function do_DATA1}
\usage{
do_DATA1(INT, MZ, TIME, count, MZmin, MZmax, MZP)
}
\arguments{
\item{INT}{}

\item{MZ}{}

\item{TIME}{}

\item{count}{}

\item{MZmin}{}

\item{MZmax}{}

\item{MZP}{}
}
\value{
DATA
}
\description{
Function do_DATA1
This funcition is used to arrange the initial data table after import
}

