% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_cdf}
\alias{read_cdf}
\title{Function read_cdf}
\usage{
read_cdf(projectpath, filepath, method = 1)
}
\arguments{
\item{projectpath}{}

\item{filepath}{}

\item{method}{}
}
\value{
Xbc, SCAN_INFO, SCAN_RANGE, file
}
\description{
Function read_cdf
}

