% Generated by roxygen2 (4.0.1): do not edit by hand
\name{replot_align}
\alias{replot_align}
\title{Function replot_align}
\usage{
replot_align(DATA, ADATA, start, stop, zoomwidth = c(0, 0), datasource,
  maintext, targetfile)
}
\arguments{
\item{DATA}{}

\item{ADATA}{}

\item{start}{}

\item{stop}{}

\item{zoomwidth}{}

\item{datasource}{}

\item{maintext}{}

\item{targetfil}{}
}
\description{
Function replot_align
}

