% Generated by roxygen2 (4.0.1): do not edit by hand
\name{replot_checkdata}
\alias{replot_checkdata}
\title{Function replot_checkdata}
\usage{
replot_checkdata(x, y, scale_text, maintext, OBS, ssX, s, coords = NULL)
}
\arguments{
\item{x}{}

\item{y}{}

\item{scale_text}{}

\item{maintext}{}

\item{OBS}{}

\item{ssX}{}

\item{s}{}

\item{coords}{}
}
\description{
Function replot_checkdata
}

