% Generated by roxygen2 (4.0.1): do not edit by hand
\name{replot_mcr}
\alias{replot_mcr}
\title{Function replot_mcr}
\usage{
replot_mcr(x, y, maintext, incl, excl, numfiles, set1, set2,
  seltype = (!missing(set1) & !missing(set2)), plotlabels = TRUE, pred)
}
\arguments{
\item{x}{}

\item{y}{}

\item{maintext}{}

\item{incl}{}

\item{excl}{}

\item{numfiles}{}

\item{set1}{}

\item{set2}{}

\item{seltype}{}

\item{plotlabels}{}

\item{pred}{}
}
\description{
Function replot_mcr
}

