% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_proc.R
\name{separate_windows}
\alias{separate_windows}
\title{Function separate_windows}
\usage{
separate_windows(projectpath, windowlist, files, edges, model, prediction)
}
\description{
Function separate_windows
}

